# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Project & Task API",
        "description": "API RESTful para la gestión de proyectos y tareas",
        "contact": {
            "email": "ecuadra06@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Listar tareas con filtros dinámicos",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (pending, in_progress, done)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "done"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrar por prioridad (low, medium, high)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high"
                            ]
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Filtrar por fecha de vencimiento (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filtrar por ID de proyecto (UUID)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtener detalles de una tarea",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la tarea",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea existente",
                "operationId": "c36c959b634c99e1e02ad2f3f2c36a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "d81cc7975ea62176b4ef8f74ed95168d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarea eliminada exitosamente"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Task": {
                "required": [
                    "id",
                    "title",
                    "status",
                    "priority",
                    "project_id"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "abc12345-6789-4def-0123-456789abcdef"
                    },
                    "title": {
                        "type": "string",
                        "example": "Preparar documentación"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción detallada de la tarea",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "done"
                        ],
                        "example": "pending"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "medium"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15",
                        "nullable": true
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T14:45:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T15:00:00Z"
                    }
                },
                "type": "object"
            },
            "TaskCreateRequest": {
                "required": [
                    "title",
                    "status",
                    "priority",
                    "project_id"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Preparar documentación"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción detallada de la tarea",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "done"
                        ],
                        "example": "pending"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "medium"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15",
                        "nullable": true
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                },
                "type": "object"
            },
            "TaskUpdateRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Actualizar documentación"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción actualizada de la tarea",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "done"
                        ],
                        "example": "in_progress"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "high"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-20",
                        "nullable": true
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                },
                "type": "object"
            }
        }
    },
...
}